# docs/openapi.yaml
openapi: 3.0.2
info:
  title: recipe-collection API
  version: 1.1.0
  description: backend service for recipe collection app
servers:
  - url: http://127.0.0.1:5000/api/1
    description: local development server
...
paths:
...
  /users/{user_id}:
    get:
      summary: get user
      description: get user by id
      operationId: get_user
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            format: int32
          required: true
          description: user id
      responses:
        '200':
          description: user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: update user
      description: update user by id
      operationId: update_user
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            format: int32
          required: true
          description: user id
      requestBody:
        description: user data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/me:
    get:
      summary: get current user
      description: get current user
      operationId: get_current_user
      tags:
        - users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users:
    get:
      summary: get users
      description: get all users (admin only)
      operationId: get_users
      tags:
        - users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: create user
      description: create user
      operationId: create_user
      tags:
        - users
      requestBody:
        description: user data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':  
          $ref: '#/components/responses/BadRequestError'
  /users/search/match:
    get:
      summary: search users
      description: search users by username or email
      operationId: search_user
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: username
          schema:
            type: string
        - in: query
          name: email
          schema:
            type: string
      responses:
        '200':
          description: users
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/exists:
    get:
      summary: check if user exists
      description: check if user exists by username or email
      operationId: search_user_exists
      tags:
        - users
      parameters:
        - in: query
          name: username
          schema:
            type: string
        - in: query
          name: email
          schema:
            type: string
      responses:
        '200':
          description: user exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
  /user_groups:
    post:
      summary: create user group
      description: create a new user group, the creating user becomes the group admin
      operationId: create_new_user_group
      tags:
        - users
      security:
        - bearerAuth: []
      requestBody:
        description: user group data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserGroup'
      responses:
        '201':
          description: user group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroup'
        '400':  
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /user_groups/{id}:
    get:
      summary: get user group
      description: get user group by id
      operationId: get_user_group
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int32
          required: true
          description: user group id
      responses:
        '200':
          description: user group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroup'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: delete user group
      description: delete user group by id, only the group admin can delete the group
      operationId: delete_user_group
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int32
          required: true
          description: user group id
      responses:
        '204':
          description: user group deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /user_groups/{id}/users:
    put:
      summary: add user to user group
      description: add user to user group by id, only the group admin can add users
      operationId: add_user_to_group
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int32
          required: true
          description: user group id
      requestBody:
        description: user data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  format: int32
      responses:
        '200':
          description: user group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroup'
        '400':  
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /user_groups/{id}/users/email:
    put:
      summary: add user to user group by email
      description: add user to user group by email, only the group admin can add users
      operationId: add_user_to_group_by_email
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int32
          required: true
          description: user group id
      requestBody:
        description: user data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        '200':
          description: user group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroup'
        '400':  
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /user_groups/{id}/users/{user_id}:
    delete:
      summary: remove user from user group
      description: remove user from user group by id, only the group admin can remove users
      operationId: remove_user_from_group
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int32
          required: true
          description: user group id
        - in: path
          name: user_id
          schema:
            type: integer
            format: int32
          required: true
          description: user id
      responses:
        '204':
          description: user removed from group
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: string
  responses:
    BadRequestError:
      description: bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: bad request
    UnauthorizedError:
      description: not authorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: not authorized
    ForbiddenError:
      description: forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Access Frobidden
    NotFoundError:
      description: not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Resource not found
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
          readOnly: true
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          readOnly: true
          items:
            type: string
            anyOf:
              - enum:
                  - admin
                  - user
        _links:
          type: object
          readOnly: true
          properties:
            self:
              type: string
              description: link to user
    UserInput:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - username
        - email
        - password
...
    UserGroup:
      type: object
      properties:
        id:
          type: integer
          format: int32
          readOnly: true
        group_name:
          type: string
        group_admin:
          type: integer
          format: int32
          readOnly: true
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
          readOnly: true
        _links:
          type: object
          readOnly: true
          properties:
            self:
              type: string
              description: link to user group
            users:
              type: array
              items:
                type: string
                format: uri
                description: link to users